Extract RegisterGuildSection
