+++ components/Account/GuildPublish.vue
@@ -2,11 +2,24 @@
   <div
     class="flex items-baseline justify-between"
   >
-    <BaseInformationCard>
+    <InformationCard>
       N'oubliez pas de cliquer sur publier une fois les informations de votre guilde renseignées.
-    </BaseInformationCard>
-    <BasePrimaryButton @click="$emit('publish')">
+    </InformationCard>
+    <PrimaryButton @click="$emit('publish')">
       Publier
-    </BasePrimaryButton>
+    </PrimaryButton>
   </div>
 </template>
+
+<script>
+import { defineComponent } from '@vue/composition-api'
+import InformationCard from '~/components/atoms/InformationCard.vue'
+import PrimaryButton from '~/components/Base/PrimaryButton.vue'
+
+export default defineComponent({
+  components: {
+    InformationCard,
+    PrimaryButton
+  }
+})
+</script>
