+++ pages/admin/guildes/_slug.vue
@@ -5,7 +5,7 @@
     </BaseHeader2>
     <UiLoader v-if="fetchState.pending" class="mx-auto" />
     <div v-else class="space-y-8">
-      <BaseInformationCard>
+      <InformationCard>
         <p v-if="guild.ownerUid">
           Le GM actuel est <span class="font-semibold">{{ guild.ownerUid }}</span>.
         </p>
@@ -15,7 +15,7 @@
         <p>
           Vous pouvez choisir un nouveau GM pour <span class="font-semibold">{{ guild.name }}</span>.
         </p>
-      </BaseInformationCard>
+      </InformationCard>
       <AdminUserInput v-model="newOwnerUid" />
       <BasePrimaryButton @click="onSave">
         Valider
@@ -27,7 +27,7 @@
 <script>
 import { defineComponent, ref, useFetch, useRoute } from '@nuxtjs/composition-api'
 import BaseHeader2 from '~/components/Base/Header2.vue'
-import BaseInformationCard from '~/components/Base/InformationCard.vue'
+import InformationCard from '~/components/atoms/InformationCard.vue'
 import BasePrimaryButton from '~/components/Base/PrimaryButton.vue'
 import AdminUserInput from '~/components/Admin/UserInput.vue'
 import useGuilds from '~/composables/database/useGuilds'
@@ -36,7 +36,7 @@ export default defineComponent({
   layout: 'admin',
   components: {
     BaseHeader2,
-    BaseInformationCard,
+    InformationCard,
     BasePrimaryButton,
     AdminUserInput
   },
