+++ pages/compte/parametres.vue
@@ -1,22 +1,6 @@
 <template>
   <div class="space-y-10">
-    <BaseInformationCard
-      v-if="!isAGuildOwner"
-    >
-      <div class="flex font-semibold mb-4 space-x-4">
-        <div>ℹ️</div>
-        <div>
-          Comment enregistrer ma guilde ?
-        </div>
-      </div>
-      <p class="mb-2">
-        Pour activer votre page de guilde, effectuez une demande auprès des <span class="font-semibold">modérateurs du Discord Sulfuron</span> en fournissant votre :
-      </p>
-      <ul class="list-disc list-inside space-y-1">
-        <li>BattleTag&trade; <em>(avec les majuscules, si nécessaire)</em></li>
-        <li>Nom de guilde</li>
-      </ul>
-    </BaseInformationCard>
+    <RegisterGuildSection />
     <BaseHeader2>
       Battle.net
     </BaseHeader2>
@@ -38,10 +22,14 @@
 
 <script>
 import { mapGetters } from 'vuex'
+import RegisterGuildSection from '~/components/organisms/Account/RegisterGuildSection.vue'
 
 export default {
   name: 'Settings',
   layout: 'account',
+  components: {
+    RegisterGuildSection
+  },
   computed: {
     ...mapGetters('account', [
       'username',
